remove working

+++ src/main/java/com/jmakulec/osm/demo/service/Patient/PatientService.java
@@ -1,6 +1,7 @@
 package com.jmakulec.osm.demo.service.Patient;
 
 import com.jmakulec.osm.demo.controller.patient.PatientRequest;
+import com.jmakulec.osm.demo.controller.patient.PatientResponse;
 import com.jmakulec.osm.demo.domain.Exam;
 import com.jmakulec.osm.demo.domain.Patient;
 import com.jmakulec.osm.demo.dto.PatientDetailDTO;
@@ -40,7 +41,29 @@ public class PatientService {
         return new PatientDetailDTO(patient);
     }
 
+    @Transactional
+    public Patient edit(PatientResponse response, Long id) throws ObjectNotFoundException{
+        return repository.findById(id)
+                .map(patient -> {
+                    patient.setName(response.getName());
+                    patient.setSurname(response.getSurname());
+                    patient.setHeightset(response.getHeightset());
+                    return repository.save(patient);
+                })
+                .orElseGet(() -> {
+                    response.setId(id);
+                    Patient repPatient = createPatientUpdate(response);
+                    return repository.save(repPatient);
+                });
+    }
 
+    public Patient createPatientUpdate(PatientResponse request) throws ObjectNotFoundException{
+        Patient patient = new Patient();
+        patient.setName(request.getName());
+        patient.setSurname(request.getSurname());
+        patient.setHeightset(request.getHeightset());
+        return patient;
+    }
 
     public Patient createPatientFromRequest(PatientRequest request) throws ObjectNotFoundException{
         Patient patient = new Patient();
@@ -50,11 +73,14 @@ public class PatientService {
         return patient;
     }
 
+    @Transactional
     public void deletePatient(long idPatient) throws ObjectNotFoundException{
+        if(examRepository.findByPatientId(idPatient) != null
+        ){
         List<Exam> list = examRepository.findByPatientIdOrderByDateAsc(idPatient);
         for (Exam alist:list) {
             examRepository.delete(alist);
-        }
+        }}
         repository.delete(repository.getOne(idPatient));
     }
 }
